add new main-page, create new-index-component with link, change appearance of result
