+++ components/PersonalAnswer/index.js
@@ -17,7 +17,9 @@ export default function PersonalAnswer({
   additionalToday,
   additionalLife,
 }) {
-  return (
+  const acceptableValue = year >= 0 && today >= 0 && life >= 0;
+
+  return acceptableValue ? (
     <>
       <StyledPersonalAnswer>
         <p>{personalAnswerText}</p>
@@ -48,5 +50,5 @@ export default function PersonalAnswer({
         </p>
       </StyledPersonalAnswer>
     </>
-  );
+  ) : null;
 }
